requirements.txt
fastapi
uvicorn
pycryptodome
requests
python-dotenv
python-multipart
inquirer